refactored to match json response edit
