+++ src/app/stores/page.tsx
@@ -13,19 +13,11 @@ export default function StoresPage() {
 
   const [expandedChainId, setExpandedChainId] = React.useState<number | null>(null)
   const [chains, setChains] = React.useState<Chain[]>([])
-  const [stores, setStores] = React.useState<Store[]>([])
 
   // Fetch chains from database on component mount
   React.useEffect(() => {
     fetchChains();
   }, [])
-
-  // Fetch stores when chains change
-  React.useEffect(() => {
-    for (const chain of chains) {
-      fetchStores(chain.id);
-    }
-  }, [chains])
 
   // Function to fetch chains from the backend API
   const fetchChains = async () => {
@@ -39,27 +31,6 @@ export default function StoresPage() {
       console.log("Chains fetched successfully:", fetchedChains);
     } catch (error) {
       console.error("Error fetching chains:", error);
-    }
-  };
-
-  const fetchStores = async (chainId: number) => {
-    try {
-      const response = await backend.get(`/chains/${chainId}/stores`);
-      const fetchedStores: Store[] = response.data.stores.map((store: any) => ({
-        id: store.ID,
-        chainId: store.chainID,
-        houseNumber: store.houseNumber,
-        street: store.street,
-        city: store.city,
-        state: store.state,
-        postCode: store.postCode,
-        country: store.country
-      }));
-      // Append new stores
-      setStores(prev => [...prev, ...fetchedStores]);
-      console.log("Stores fetched successfully:", fetchedStores);
-    } catch (error) {
-      console.error("Error fetching stores:", error);
     }
   };
 
@@ -79,7 +50,7 @@ export default function StoresPage() {
       <main>
         <h1>Stores Page</h1>
         <ChainsPanel chains={chains} expandedChainId={expandedChainId} setExpandedChainId={setExpandedChainId} onAddChain={addChain} />
-        <StoresPanel chains={chains} stores={stores} expandedChainId={expandedChainId} onAddStore={addStore}/>
+        <StoresPanel chains={chains} expandedChainId={expandedChainId} onAddStore={addStore}/>
       </main>
     </div>
   )
